chore: seperate init database

+++ backend/app.py
@@ -4,7 +4,7 @@ from db.db import db
 from routes.ping import health_bp
 from routes.doctor import doctor_bp
 
-def create_app():
+def create_app(init_db=False):
     app = Flask(__name__)
     app.config.from_object(Config)
 
@@ -12,8 +12,9 @@ def create_app():
 
     with app.app_context():
         import models
-        db.drop_all()
-        db.create_all()
+        if init_db:
+            db.drop_all()
+            db.create_all()
 
     app.register_blueprint(health_bp)
     app.register_blueprint(doctor_bp)
